+++ packages/moonstone/CHANGELOG.md
@@ -10,6 +10,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Changed
 
+- `moonstone/VideoPlayer` to handle play, pause, stop, fast forward and rewind on remote controller
+
 ### Fixed
 
 - `moonstone/Scroller` not scrolled via 5 way when `moonstone/ExpandableList` is opened.